Moved `ReferenceableType` and `SharedReference` to shared_reference.dart
